+++ lib/fl_hover.dart
@@ -91,6 +91,10 @@ export 'src/models/cursor_behavior.dart';
 export 'src/components/cursor_showcase/cursor_showcase_widget.dart';
 export 'src/components/cursor_showcase/models/cursor_showcase_theme.dart';
 
+// Export the Glitch Image component
+export 'src/components/glitch_image/glitch_image_widget.dart';
+export 'src/components/glitch_image/models/glitch_image_theme.dart';
+
 // Export the Split Image component
 export 'src/components/split_image/split_image_widget.dart';
 export 'src/components/split_image/models/split_image_theme.dart';